Adding a custom hook to do AJAX calls.
