+++ src/components/todo/todo.scss
@@ -12,7 +12,6 @@ body, html {
 }
 
 
-
 div {
     .complete-false,
     .complete-true {
@@ -23,6 +22,9 @@ div {
             margin-right: 8px;
             cursor: pointer;
         }
+    }
+    .container {
+        padding-top: 20px;
     }
 }
